chore(coupon-api): move JPA and Redis components to appropriate packages
