+++ coupon/coupon-api/src/main/java/dev/be/coupon/api/coupon/infrastructure/redis/CouponCountRedisRepository.java
@@ -1,4 +1,4 @@
-package dev.be.coupon.api.coupon.infrastructure;
+package dev.be.coupon.api.coupon.infrastructure.redis;
 
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Repository;
