+++ coupon/coupon-api/src/test/java/dev/be/coupon/api/coupon/infrastructure/redis/CouponCountRedisRepositoryTest.java
@@ -1,4 +1,4 @@
-package dev.be.coupon.api.coupon.infrastructure;
+package dev.be.coupon.api.coupon.infrastructure.redis;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.jupiter.api.DisplayName;
